chore: Refactored out QueryParams to a separate dir.
