+++ app/api/api_v1/endpoints/bodies.py
@@ -1,44 +1,14 @@
 from typing import Any, Optional
 
-from fastapi import APIRouter, Depends, Query, Request
-from pydantic import BaseModel
+from fastapi import APIRouter, Depends, Request
 from sqlalchemy.orm import Session
 
 from app import crud, schemas
 from app.api import deps
+from app.api.api_v1.params.bodies import BodyQueryParams
 from app.api.paginator import PaginatedResponse
 
 router = APIRouter()
-
-
-class BodyQueryParams(BaseModel):
-    limit: Optional[int] = Query(
-        default=20, title="The number of records to return per page", deprecated=True
-    )
-
-    ra: Optional[float] = Query(
-        default=None, title="The right ascension to search", deprecated=True
-    )
-
-    dec: Optional[float] = Query(
-        default=None, title="The declination to search", deprecated=True
-    )
-
-    radius: Optional[float] = Query(
-        default=None,
-        title="The search radius from the point { ra, dec }",
-        deprecated=True,
-    )
-
-    name: Optional[str] = Query(
-        default=None, title="The name of the body object to search", deprecated=True
-    )
-
-    constellation: Optional[str] = Query(
-        default=None,
-        title="The constellation of the body object to search",
-        deprecated=True,
-    )
 
 
 @router.get("/", name="bodies:list", response_model=PaginatedResponse[schemas.Body])
